Nonparametric bias-corrected and accelerated bootstrap confidence limits

bcajack computes nonparametric confidence intervals for bootstrap estimates. For reproducibility, save or set the random number state before calling this routine.

bcajack(x, B, func, ..., m = nrow(x), mr = 5, K = 2, J = 10,
  alpha = c(0.025, 0.05, 0.1, 0.16, 0.5, 0.84, 0.9, 0.95, 0.975),
  verbose = TRUE, sw = 0)

Arguments

x
an nxp data matrix, rows are observed p-vectors, assumed to be independently sampled from target population. If p is 1 then x can be a vector.

B
number of bootstrap replications. 'B' can also be a vector of B bootstrap replications of the estimated parameter of interest, computed separately.

func
function \(\hat{\theta}\)=func(x) computing estimate of the parameter of interest; func(x) should return a real value for any n' x p matrix x', n' not necessarily equal to n.

...
additional arguments for func.

m
integer m <= n; collects the n rows of x into m groups to speed up the jackknife calculations for estimating the acceleration value 'a'; typically m=20 or 40; does not have to exactly divide n.

mr
if m < n then mr repetions of the randomly grouped jackknife calculations are averaged.

K
If K > 0, bcajack also returns estimates of 'internal standard error', that is, of the variability due to stopping at B bootstrap replications rather than going on to infinity. These are obtained from a second type of jackknifing, taking an average of K separate jackknife estimates, each randomly splitting the B bootstrap replications into J groups.

J
the number of groups into which the bootstrap replications are split

alpha
percentiles desired for the bca confidence limits.

verbose
logical for verbose messages

sw
switch that controls output, eg sw=5 returns the B bootstrap replications as well as the bca output.

rou
rounding parameter for the output.

Value

a named list of several items:

lims

first column shows the estimated bca confidence limits at the requested alpha percentiles. These can be compared with the standard limits thetahat+sdboot*z[alpha], third column. The second column "jacksd" gives the internal standard errors for the bca limits, quite small in the example. Column 4, "pct", gives the percentiles of the ordered B bootstrap replications corresponding to the bca limits, eg the 897th largest replication equalling the .975 bca limit .557.

stats

top line of stats shows 5 estimates: thet is func(x), original point estimate of the parameter of interest; sdboot is its bootstrap estimate of standard error; z0 is the bca bias correction value, in this case quite negative; a is the "acceleration", a component of the bca limits (nearly zero here); sdjack is the jackknife estimate of standard error for thet. Bottom line gives the internal standard errors for the five quantities above. This is substantial for z0 above.

B.mean

bootstrap sample size B, and the mean of the B bootstrap replications theathat*.

Details

Bootstrap confidence intervals depend on three elements:

the cdf of the B bootstrap replications t[i]*, i=1...B

the bias-correction number z0=qnorm(#{t[i]* < t0}/B ) where t0=func(x) is the original estimate

the acceleration number \(a\) that measures the rate of change in sd{t0} as x, the data changes.

The first two of these depend only on the bootstrap distribution, and not how it is generated: parametrically or non-parametrically. Program bcajack can be used in a hybrid fashion in which the vector tt of B bootstrap replications is first generated from a parametric model.

So, in the diabetes example below, we might first draw bootstrap samples \(y^* \sim N(X\hat{\beta}, \hat{\sigma}^2I)\) where \(\hat{beta}\) and \(\hat{sigma}\) were obtained from lm(y~X); each \(y^*\) would then provide a bootstrap replication t* = rfun(cbind(X,y*)). Then we could get bca intervals from bcajack(Xy, tt, rfun ....) with tt, the vector of B t* values. The only difference from a full parametric bca analysis would lie in the nonparametric estimation of a, often a negligible error.

References

DiCiccio T and Efron B (1996). Bootstrap confidence intervals. Statistical Science 11, 189-228

Efron B (1987). Better bootstrap confidence intervals. JASA 82 171-200

Examples

data(diabetes, package = "lars")
Xy <- cbind(diabetes$x, diabetes$y)
rfun <- function(Xy) {
  y <- Xy[, 11]
  X <- Xy[, 1:10]
  summary(lm(y~X) )$adj.r.squared
}
bcajack(x = Xy, B = 1000, func = rfun, m = 40)

#>   |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%

#> Warning: longer object length is not a multiple of shorter object length

#> $call
#> bcajack(x = Xy, B = 1000, func = rfun, m = 40)
#> 
#> $lims
#>         bcalims      jacksd  standard   pct
#> 0.025 0.4285596 0.003637365 0.4456986 0.005
#> 0.05  0.4396789 0.003266602 0.4554836 0.012
#> 0.1   0.4526139 0.004371316 0.4667650 0.030
#> 0.16  0.4674165 0.003120873 0.4756799 0.056
#> 0.5   0.4974291 0.001610290 0.5065603 0.282
#> 0.84  0.5282229 0.002728307 0.5374407 0.661
#> 0.9   0.5386519 0.002827792 0.5463557 0.757
#> 0.95  0.5489462 0.002864281 0.5576371 0.854
#> 0.975 0.5561995 0.002685774 0.5674220 0.913
#> 
#> $stats
#>          thet       sdboot          z0            a     sdjack
#> est 0.5065603 0.0310524656 -0.28714669 -0.007570513 0.03370235
#> jsd 0.0000000 0.0008317929  0.04058162  0.000000000 0.00000000
#> 
#> $B.mean
#> [1] 1000.0000000    0.5156306
#> 
#> $ustats
#>      ustat        sdu 
#> 0.49749003 0.04053599 
#>