Nonparametric bias-corrected and accelerated bootstrap confidence limits

bcajack2 is a version of bcajack that allows all the recomputations of the original statistic function func to be carried out separately. This is an advantage if func is time-consuming, in which case the B replications for the nonparametric bca calculations might need to be done on a distributed basis.

To use bcajack2 in this mode, we first compute a list Blist via Blist <- list(Y = Y,tt = tt,t0 = t0). Here tt is a vector of length B having i-th entry tt[i] <- func(x[Ii,], ...), where x is the n x p data matrix and Ii <- sample(n,n,T), a bootstrap vector of indices. Y is a B x n "count" matrix, whose i-th row is the counts corresponding to Ii. For example if n = 5 and Ii=(2,5,2,1,4), then Yi = (1,2,0,1,1). Having computed Blist, bcajack2 is invoked as bcajack2(Blist) without need to enter the function func.

bcajack2(x, B, func, ..., m = nrow(x), mr, pct = 0.333, K = 2,
  J = 12, alpha = c(0.025, 0.05, 0.1, 0.16, 0.5, 0.84, 0.9, 0.95,
  0.975), verbose = TRUE, sw = 0)

Arguments

x
an nxp data matrix, rows are observed p-vectors, assumed to be independently sampled from target population. If p is 1 then x can be a vector.

B
number of bootstrap replications. 'B' can also be a vector of B bootstrap replications of the estimated parameter of interest, computed separately. If B is Blist as explained above, x is not needed.

func
function \(\hat{\theta}\)=func(x) computing estimate of the parameter of interest; func(x) should return a real value for any n' x p matrix x', n' not necessarily equal to n.

...
additional arguments for func.

m
integer m <= n; collects the n rows of x into m groups to speed up the jackknife calculations for estimating the acceleration value 'a'; typically m=20 or 40; does not have to exactly divide n.

mr
if m < n then mr repetions of the randomly grouped jackknife calculations are averaged.

K
If K > 0, bcajack also returns estimates of 'internal standard error', that is, of the variability due to stopping at B bootstrap replications rather than going on to infinity. These are obtained from a second type of jackknifing, taking an average of K separate jackknife estimates, each randomly splitting the B bootstrap replications into J groups.

J
the number of groups into which the bootstrap replications are split

alpha
percentiles desired for the bca confidence limits.

verbose
logical for verbose messages

sw
switch that controls output, eg sw=5 returns the B bootstrap replications as well as the bca output.